Create cross_validation().py
Being a novice to ML, one of the first challenges we face is painting the right picture of the dataset we are given to play with. 
Data are not visible, they are all numbers, but we still have to 'see' them to understand what they are. 

I couldn't understand the code for "Implementing Cross-Validation" box in Chapter 3: Classification in the book by Mr. Geron.
So I built upon his code to make the data arrays more visible. 

This function prints some more visible information about the MNIST dataset so we can have a better understanding of what it is.

Reference
Aurélien Géron (2017) Hands-On Machine Learning with Scikit-Learn & TensorFlow (p. 83)