Create mnist_y_info().py
Being a novice to ML, one of the first challenges we face is painting the right picture of the dataset we are given to play with. 
Data are not visible, they are all numbers, but we still have to 'see' them to understand what they are. 

MNIST is an extremely complicated dataset for machine learning newbies. 
I couldn't grasp the idea of MNIST so I wrote a simple function that returns an overview of the y labels of MNIST.